fix: implement FormWrapper on ScenarioDetail>Flashcards
