+++ frontend/src/organisms/Scenario/ScenarioDetail.tsx
@@ -1,5 +1,6 @@
 import { Button } from "@/atoms/Button";
 import EditableMarkdown from "@/atoms/EditableMarkdown";
+import FormWrapper from "@/atoms/FormWrapper";
 import Markdown from "@/atoms/Markdown";
 import { useDeleteScenarioMutation } from "@/lib/graphql/generated/graphql-types";
 import { Input } from "@/lib/shadcn/generated/ui/input";
@@ -27,9 +28,13 @@ type Props = {
   scenario: Entities.Scenario;
 };
 export default function ScenarioDetail({ scenario }: Props) {
+  const [locked, setLocked] = useState<boolean>(true);
+  const toggleLocked = () => {
+    setLocked(!locked);
+  };
+
   const [needle, setNeedle] = useState<string>("");
   const [currPlan, setCurrPlan] = useState<Entities.Plan>(scenario.plans[0]);
-  const [isEditing, setIsEditing] = useState<boolean>(false);
   const currentUser = useCurrentUser();
   const [deleteScenario] = useDeleteScenarioMutation();
   const navigate = useNavigate();
@@ -49,18 +54,6 @@ export default function ScenarioDetail({ scenario }: Props) {
       navigate("/scenarios");
     }
   };
-  const hEnterEditMode = async () => {
-    setIsEditing(true);
-  };
-
-  const hCancelEditMode = async () => {
-    setIsEditing(false);
-  };
-  const hUpdateScenario = async () => {
-    console.log("TODO!");
-    setIsEditing(false);
-    // TODO: implement
-  };
 
   return (
     <>
@@ -75,11 +68,11 @@ export default function ScenarioDetail({ scenario }: Props) {
           )}
         </TabsList>
         <TabsContent value="home">
-          {isEditing ? (
-            <EditableMarkdown source={scenario.fullStory} />
-          ) : (
-            <Markdown value={scenario.fullStory} />
-          )}
+          <FormWrapper
+            baseComp={<Markdown value={scenario.fullStory} />}
+            formComp={<EditableMarkdown source={scenario.fullStory} />}
+            locked={locked}
+          />
         </TabsContent>
         <TabsContent value="plans">
           {currPlan && (
@@ -111,7 +104,7 @@ export default function ScenarioDetail({ scenario }: Props) {
             data={scenario.flashcards.filter((fcard) =>
               fcard.title.toLowerCase().includes(needle.toLowerCase()),
             )}
-            isEditing={isEditing}
+            isEditing={!locked}
           />
         </TabsContent>
         {currentUser?.id === scenario.owner.id && (
@@ -121,20 +114,9 @@ export default function ScenarioDetail({ scenario }: Props) {
             <Button variant={"destructive"} onClick={hDeleteScenario}>
               Delete
             </Button>
-            {!isEditing ? (
-              <Button variant={"secondary"} onClick={hEnterEditMode}>
-                Edit
-              </Button>
-            ) : (
-              <>
-                <Button type="reset" onClick={hCancelEditMode}>
-                  Cancel
-                </Button>
-                <Button type="button" onClick={hUpdateScenario}>
-                  Confirm
-                </Button>
-              </>
-            )}
+            <Button onClick={toggleLocked}>
+              {locked ? "Unlock" : "Relock"}
+            </Button>
           </TabsContent>
         )}
       </Tabs>
