+++ frontend/src/pages/public/SandboxPage.tsx
@@ -1,16 +1,10 @@
 import { Button } from "@/atoms/Button";
 import EditableMarkdown from "@/atoms/EditableMarkdown";
+import FormWrapper from "@/atoms/FormWrapper";
 import Markdown from "@/atoms/Markdown";
-import { useEventListener } from "@/lib/hooks/useEventListener";
 import FlashCard from "@/organisms/Flashcard/FlashCard";
 import FlashCardForm from "@/organisms/Flashcard/FlashCardForm";
-import {
-  type ReactElement,
-  cloneElement,
-  useEffect,
-  useId,
-  useState,
-} from "react";
+import { useState } from "react";
 
 const scenario = {
   title: "A la chasse aux gobs",
@@ -103,38 +97,5 @@ export default function SandboxPage() {
         locked={locked}
       />
     </>
-  );
-}
-
-type FormWrapperProps = {
-  baseComp: ReactElement;
-  formComp: ReactElement<{ formCompId: string }>;
-  locked: boolean;
-};
-function FormWrapper({ baseComp, formComp, locked }: FormWrapperProps) {
-  const [editable, setEditable] = useState<boolean>(locked);
-  const id = useId();
-  const toggleEditable = () => {
-    setEditable(!editable);
-  };
-  useEventListener("FormWrapper-submit-child", (detail) => {
-    if (detail.uuid === id) setEditable(false);
-  });
-
-  useEffect(() => {
-    if (locked) setEditable(false);
-  }, [locked]);
-
-  return (
-    <div className="relative">
-      {editable
-        ? cloneElement(formComp, { formCompId: id })
-        : cloneElement(baseComp)}
-      {!locked && (
-        <div className="absolute right-0 bottom-0">
-          {!editable && <Button onClick={toggleEditable}>✏️</Button>}
-        </div>
-      )}
-    </div>
   );
 }
